Time: 26 ms (55.42%), Space: 5.8 MB (72.63%) - LeetHub

+++ 2240-number-of-ways-to-buy-pens-and-pencils/2240-number-of-ways-to-buy-pens-and-pencils.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     long long waysToBuyPensPencils(int total, int cost1, int cost2) {
-        long long res=0;
-        for(long long int i=0 ;i*cost1<=total ;i++){
+        long long int res=0;
+        for(int i=0 ;i*cost1<=total ;i++){
             res+=(total-i*cost1)/cost2+1;
         }
         return res;
